đóng gói minimax ai thành ai1. thêm sắp xếp các nước đi trong hàm minimax
